% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LED.R
\docType{class}
\name{LED-class}
\alias{LED-class}
\title{LED Class Representation.}
\description{
LED Class Representation.
}
\section{Slots}{

\describe{
\item{\code{pin}}{The led's assigned pin.}

\item{\code{led}}{The internal representation of the led.}
}}

