% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_led_value.R
\name{set_led_value}
\alias{set_led_value}
\alias{set_led_value,LED-method}
\title{Set the value of a led instance.}
\usage{
set_led_value(x, value)

\S4method{set_led_value}{LED}(x, value)
}
\arguments{
\item{x}{A LED instance.}

\item{value}{A logical of length one. If `FALSE`, the LED will be switched
off. If `TRUE`, the LED will be switched on.}
}
\value{
The same LED instance provided as input.
}
\description{
Sets the value of a given led instance.
}
\section{Methods (by class)}{
\itemize{
\item \code{LED}: 
}}

\examples{
\dontrun{
# Get a led instance.
led <- get_led(14)
# Turn the led on.
set_led_value(led, TRUE)
# Turn the led off.
set_led_value(led, FALSE)
}

}
