% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_led_value.R
\name{get_led_value}
\alias{get_led_value}
\alias{get_led_value,LED-method}
\title{Get the value of a led instance.}
\usage{
get_led_value(x)

\S4method{get_led_value}{LED}(x)
}
\arguments{
\item{x}{A LED instance.}
}
\value{
A logical of length one with the actual value of the led.
}
\description{
Gets the value of a given led instance.
}
\section{Methods (by class)}{
\itemize{
\item \code{LED}: 
}}

\examples{
\dontrun{
# Get a led instance with initial value TRUE.
led <- get_led(14, TRUE)
# Read the value of the led (should be TRUE).
get_led_value(led)
}

}
