% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_led_value.R
\name{set_led_value}
\alias{set_led_value}
\alias{set_led_value,LED-method}
\title{Set the value of a led instance.}
\usage{
set_led_value(x, value)

\S4method{set_led_value}{LED}(x, value)
}
\arguments{
\item{x}{A LED instance.}

\item{value}{A numeric of length one. If 0, the LED will be off. Other
values between 0 and 1 can be specified as brightness for the LED.}
}
\value{
The same LED instance provided as input.
}
\description{
Sets the value of a given led instance.
}
\section{Methods (by class)}{
\itemize{
\item \code{LED}: 
}}

\examples{
\dontrun{
# Get a led instance.
led <- get_led(14)
# Turn the led to its maximum intensity.
set_led_value(led, 1)
# Turn the led off.
set_led_value(led, 0)
}

}
